Quantize and dequantize modules
Added the quantize and dequantize operations mentioned in ticket #1 as
classes deriving from torch.nn.Module

The quantization bits can be chosen per channel of each batch element.
This requirement is not mentioned on the ticket, where I did only
specify to make it constant to 8 bits. The plan was to decide later at
what granularity we wanted to specify the quantization level, but I made
up my mind on this.

The quantization degree will be chosen per channel because that way we
can save channels as grayscale images in PNG. The PNG format supports
saving grayscale images with 1, 2, 4, 8 or 16 bits color depths, which
will perfectly suit the purpose of saving channels coded with different
quantization degrees.

Being able to use PNG for storing and loading channels gives us access
to the state of art in entropic coding almost for free.

The implemented modules (Quantize and Dequantize) are designed to be
used together. The output from Quantize matches the input from
Dequantize, so they can both be chained one after another on a model.

Together with the implementation I provide a unit test that chains
together an instance of Quantize and another of Dequantize. The test
feeds a simple handcrafted input and compares the intermediate results
and outputs against the expected values.

We need to setup something like pytest for our unit tests, but
meanwhile I forcefully invoke the test from train command line option.
Note this is a hack that will only be fixed when we have have working
prototype.