Added a class encapsulating the bit stream writing
As part of ticket #9 I looked at the options for writing grayscale
PNG files with an arbitrary number of bits per pixel. However, the
support for fancy formats like 1, 2, 4, 8 and 16 bits is only
available for paletted images in PIL. Tried also OpenCV2 where the
support was even worse, not even having the option of forcing the
number of bits by using a palette.

It then became evident to me that writting PNG files with arbitrary
number of bits, even though it is supported by the PNG spec, is
just not supported in the python libraries and going that way would
be like openning a can of worms.

As alternative I am going to write the quantized tensors as binary
files with a format defined by myself. For this task I need a bit
stream writer and a reader.

After giving a try to the bitstream python package I found it was too
complex and too broken for being useful, so I wrote my own bit stream
writer out of need.

Will also make my own reader on another patch for addressing #10 (the
reader).

When the tensor writer (#9) and the reader (#10), then I will document
the used format (#12).