Fixed a typo in the functional.pad parameters for CropsDataset
For the case of not having padding on the bottom edge I wrote the code
in a way that "bottom" was undefined, and therefore assumed to be None,
rather being a value explicitly defined as zero.

Looks like None is accepted a parameter in the padding tuple for
functional.pad and interpreted as zero. Because of this reason the code
was actually doing what I expected, but in a way that was not obvious.

This patch fixes the typo.