A dataset returning crops from the images on a given directory
This dataset feeds crops from images on a given directory. The crops
are padded with a reflection of the original image on the right and
bottom edges when the image dimensions are not a multiple of the crop size.

When no cropping is required and the image size is fixed (like in the
VoxCeleb1 dataset) then we can just pick whole image crops, which is the
same as feeding all the images from a given directory.

This patch closes ticket #4