Improved CropsDataset
The existing implementation had several issues:

- On the class constructor it was reading the raster data of all the images
  for creating padded versions. This was very slow.

- All the padded images were kept in memory, which made impossible to
  handle big datasets. As a mater of fact the previous revision
  failed to load the 110K faces we had selected from VoxCeleb1.

- There was no way to limit the amount of images read, which was a
  problem for big datasets.

I have addressed all the three issues in the new revision. While I was
at it I also made reading the image metadata for looking for the image
dimensions optional, since the case of being able to assume a given
size is going to be very common.

For the general purpose case in which you still want to read the image
dimensions I have added a progress bar.