Updated autoencoder.utils.create_dataloaders with a new dataset
The newly introduced dataset is based on VoxCeleb2, whereas the
previous one was based on VoxCeleb. However, more important than the
version upgrade is the fact that the new dataset ensures the train
and test sets have no overlapping whatsoever.

There is no overlapping  in terms of identities or utterances, which was
an issue with the dataset previously used.