Fixed a bug when loading the model for inference from the command line
The model was loaded in training mode (rather than evaluation mode), because
that was the mode in which it was saved. Then I did not switch to evaluation
mode after loading, which led to inconsistent inferences that could not
reproduce the good results seen oin TensorBoard for the test set.

Now I switch to evaluation mode right after loading the model and the results
from the command line inferences match the good quality I was achieving with
the test set during the training.